Split state.py into formatting.py, pipeline.py, and slim state.py
Extract formatting/UI code (detect_categories, format_reason,
block_and_print, fail_state) into formatting.py (~187 lines).
Extract orchestration (evaluate_changes, prepare_run_stop_checks,
compush_check, run_stop_checks) into pipeline.py (~302 lines).
Slim state.py to data structures and constants (~133 lines).

Update hook.py to import run_stop_checks from pipeline.
Update test mock targets from state_mod to pipeline_mod and
formatting_mod at call sites.

All 107 tests pass. Format, lint, typecheck all green.
