+++ tests/test_hook.py
@@ -29,8 +29,10 @@ import pytest
 
 from post_turn_quality_stop_hook import driver as driver_mod
 from post_turn_quality_stop_hook import execution as exec_mod
+from post_turn_quality_stop_hook import formatting as formatting_mod
 from post_turn_quality_stop_hook import git as git_mod
 from post_turn_quality_stop_hook import hook
+from post_turn_quality_stop_hook import pipeline as pipeline_mod
 from post_turn_quality_stop_hook import state as state_mod
 
 # ---------------------------------------------------------------------------
@@ -276,13 +278,13 @@ class TestCompushCheck:
         """Dirty tree + upstream -> block with push message."""
         with (
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(True, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(True, None)
             ),
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=("origin/feature", None)
+                pipeline_mod, "get_upstream_ref", return_value=("origin/feature", None)
             ),
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         out = json.loads(capsys.readouterr().out)
         assert out["decision"] == "block", (
@@ -296,13 +298,13 @@ class TestCompushCheck:
         """Dirty tree + no upstream -> block with fallback text."""
         with (
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(True, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(True, None)
             ),
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=(None, "no upstream")
+                pipeline_mod, "get_upstream_ref", return_value=(None, "no upstream")
             ),
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         out = json.loads(capsys.readouterr().out)
         assert out["decision"] == "block", (
@@ -316,25 +318,25 @@ class TestCompushCheck:
         """Clean tree -> no output, exit 0."""
         with (
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=("origin/feature", None)
+                pipeline_mod, "get_upstream_ref", return_value=("origin/feature", None)
             ),
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(False, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(False, None)
             ),
             mock.patch.object(
-                state_mod, "has_unpushed_commits", return_value=(False, None)
+                pipeline_mod, "has_unpushed_commits", return_value=(False, None)
             ),
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         assert capsys.readouterr().out == "", "expected no hook output for clean tree"
 
     def test_error_checking_changes(self, capsys: pytest.CaptureFixture[str]) -> None:
         """Error from has_uncommitted_changes -> silent exit 0."""
         with mock.patch.object(
-            state_mod, "has_uncommitted_changes", return_value=(None, "oops")
+            pipeline_mod, "has_uncommitted_changes", return_value=(None, "oops")
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         assert capsys.readouterr().out == "", (
             "expected no hook output when change check errors are suppressed"
@@ -346,16 +348,16 @@ class TestCompushCheck:
         """Clean tree + ahead of upstream -> block with push-only message."""
         with (
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=("origin/feature", None)
+                pipeline_mod, "get_upstream_ref", return_value=("origin/feature", None)
             ),
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(False, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(False, None)
             ),
             mock.patch.object(
-                state_mod, "has_unpushed_commits", return_value=(True, None)
+                pipeline_mod, "has_unpushed_commits", return_value=(True, None)
             ),
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         out = json.loads(capsys.readouterr().out)
         assert out["decision"] == "block", (
@@ -369,14 +371,14 @@ class TestCompushCheck:
         """Clean tree + no upstream -> no ahead check and no output."""
         with (
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=(None, "no upstream")
+                pipeline_mod, "get_upstream_ref", return_value=(None, "no upstream")
             ),
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(False, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(False, None)
             ),
-            mock.patch.object(state_mod, "has_unpushed_commits") as mock_ahead,
+            mock.patch.object(pipeline_mod, "has_unpushed_commits") as mock_ahead,
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         mock_ahead.assert_not_called()
         assert capsys.readouterr().out == "", (
@@ -389,16 +391,16 @@ class TestCompushCheck:
         """Ahead check errors stay silent to preserve hook contract."""
         with (
             mock.patch.object(
-                state_mod, "get_upstream_ref", return_value=("origin/feature", None)
+                pipeline_mod, "get_upstream_ref", return_value=("origin/feature", None)
             ),
             mock.patch.object(
-                state_mod, "has_uncommitted_changes", return_value=(False, None)
+                pipeline_mod, "has_uncommitted_changes", return_value=(False, None)
             ),
             mock.patch.object(
-                state_mod, "has_unpushed_commits", return_value=(None, "oops")
+                pipeline_mod, "has_unpushed_commits", return_value=(None, "oops")
             ),
         ):
-            rc = state_mod.compush_check(REPO)
+            rc = pipeline_mod.compush_check(REPO)
         assert rc == 0, f"expected compush_check rc 0 but got {rc!r}"
         assert capsys.readouterr().out == "", (
             "expected no hook output when ahead check errors are suppressed"
@@ -468,24 +470,26 @@ class TestRunStopChecksCompush:
     def test_compush_triggers_after_success(self) -> None:
         """compush=True + quality pass + dirty -> compush_check called."""
         with (
-            mock.patch.object(state_mod, "repo_root", return_value=(REPO, None)),
+            mock.patch.object(pipeline_mod, "repo_root", return_value=(REPO, None)),
             mock.patch.object(
-                state_mod, "ensure_base_ref", return_value=(True, None, False)
-            ),
-            mock.patch.object(state_mod, "merge_base", return_value=("abc123", None)),
-            mock.patch.object(
-                state_mod, "changed_files", return_value=(["src/foo.py"], None)
+                pipeline_mod, "ensure_base_ref", return_value=(True, None, False)
             ),
             mock.patch.object(
-                state_mod, "select_build_driver", return_value=(self.driver, None)
+                pipeline_mod, "merge_base", return_value=("abc123", None)
             ),
-            mock.patch.object(state_mod, "evaluate_changes", return_value=0),
             mock.patch.object(
-                state_mod, "compush_check", return_value=0
+                pipeline_mod, "changed_files", return_value=(["src/foo.py"], None)
+            ),
+            mock.patch.object(
+                pipeline_mod, "select_build_driver", return_value=(self.driver, None)
+            ),
+            mock.patch.object(pipeline_mod, "evaluate_changes", return_value=0),
+            mock.patch.object(
+                pipeline_mod, "compush_check", return_value=0
             ) as mock_compush,
             mock.patch("shutil.which", return_value="/usr/bin/git"),
         ):
-            rc = state_mod.run_stop_checks(
+            rc = pipeline_mod.run_stop_checks(
                 REPO,
                 "origin/main",
                 state_mod.StopCheckOptions(
@@ -500,22 +504,24 @@ class TestRunStopChecksCompush:
     def test_compush_skipped_when_disabled(self) -> None:
         """compush=False -> compush_check not called."""
         with (
-            mock.patch.object(state_mod, "repo_root", return_value=(REPO, None)),
+            mock.patch.object(pipeline_mod, "repo_root", return_value=(REPO, None)),
             mock.patch.object(
-                state_mod, "ensure_base_ref", return_value=(True, None, False)
-            ),
-            mock.patch.object(state_mod, "merge_base", return_value=("abc123", None)),
-            mock.patch.object(
-                state_mod, "changed_files", return_value=(["src/foo.py"], None)
+                pipeline_mod, "ensure_base_ref", return_value=(True, None, False)
             ),
             mock.patch.object(
-                state_mod, "select_build_driver", return_value=(self.driver, None)
+                pipeline_mod, "merge_base", return_value=("abc123", None)
             ),
-            mock.patch.object(state_mod, "evaluate_changes", return_value=0),
-            mock.patch.object(state_mod, "compush_check") as mock_compush,
+            mock.patch.object(
+                pipeline_mod, "changed_files", return_value=(["src/foo.py"], None)
+            ),
+            mock.patch.object(
+                pipeline_mod, "select_build_driver", return_value=(self.driver, None)
+            ),
+            mock.patch.object(pipeline_mod, "evaluate_changes", return_value=0),
+            mock.patch.object(pipeline_mod, "compush_check") as mock_compush,
             mock.patch("shutil.which", return_value="/usr/bin/git"),
         ):
-            state_mod.run_stop_checks(
+            pipeline_mod.run_stop_checks(
                 REPO,
                 "origin/main",
                 state_mod.StopCheckOptions(
@@ -529,22 +535,24 @@ class TestRunStopChecksCompush:
     def test_compush_skipped_on_quality_failure(self) -> None:
         """Quality check failure (nonzero rc) -> compush_check not called."""
         with (
-            mock.patch.object(state_mod, "repo_root", return_value=(REPO, None)),
+            mock.patch.object(pipeline_mod, "repo_root", return_value=(REPO, None)),
             mock.patch.object(
-                state_mod, "ensure_base_ref", return_value=(True, None, False)
-            ),
-            mock.patch.object(state_mod, "merge_base", return_value=("abc123", None)),
-            mock.patch.object(
-                state_mod, "changed_files", return_value=(["src/foo.py"], None)
+                pipeline_mod, "ensure_base_ref", return_value=(True, None, False)
             ),
             mock.patch.object(
-                state_mod, "select_build_driver", return_value=(self.driver, None)
+                pipeline_mod, "merge_base", return_value=("abc123", None)
             ),
-            mock.patch.object(state_mod, "evaluate_changes", return_value=1),
-            mock.patch.object(state_mod, "compush_check") as mock_compush,
+            mock.patch.object(
+                pipeline_mod, "changed_files", return_value=(["src/foo.py"], None)
+            ),
+            mock.patch.object(
+                pipeline_mod, "select_build_driver", return_value=(self.driver, None)
+            ),
+            mock.patch.object(pipeline_mod, "evaluate_changes", return_value=1),
+            mock.patch.object(pipeline_mod, "compush_check") as mock_compush,
             mock.patch("shutil.which", return_value="/usr/bin/git"),
         ):
-            rc = state_mod.run_stop_checks(
+            rc = pipeline_mod.run_stop_checks(
                 REPO,
                 "origin/main",
                 state_mod.StopCheckOptions(
@@ -559,19 +567,21 @@ class TestRunStopChecksCompush:
     def test_compush_runs_when_no_files_changed(self) -> None:
         """compush=True still runs when there are no changed files to lint."""
         with (
-            mock.patch.object(state_mod, "repo_root", return_value=(REPO, None)),
+            mock.patch.object(pipeline_mod, "repo_root", return_value=(REPO, None)),
             mock.patch.object(
-                state_mod, "ensure_base_ref", return_value=(True, None, False)
+                pipeline_mod, "ensure_base_ref", return_value=(True, None, False)
             ),
-            mock.patch.object(state_mod, "merge_base", return_value=("abc123", None)),
-            mock.patch.object(state_mod, "changed_files", return_value=([], None)),
-            mock.patch.object(state_mod, "evaluate_changes") as mock_evaluate,
             mock.patch.object(
-                state_mod, "compush_check", return_value=0
+                pipeline_mod, "merge_base", return_value=("abc123", None)
+            ),
+            mock.patch.object(pipeline_mod, "changed_files", return_value=([], None)),
+            mock.patch.object(pipeline_mod, "evaluate_changes") as mock_evaluate,
+            mock.patch.object(
+                pipeline_mod, "compush_check", return_value=0
             ) as mock_compush,
             mock.patch("shutil.which", return_value="/usr/bin/git"),
         ):
-            rc = state_mod.run_stop_checks(
+            rc = pipeline_mod.run_stop_checks(
                 REPO,
                 "origin/main",
                 state_mod.StopCheckOptions(
@@ -610,7 +620,7 @@ class TestRunOSError:
     ) -> None:
         """Full pipeline exits cleanly when start_cwd does not exist."""
         with mock.patch("shutil.which", return_value="/usr/bin/git"):
-            rc = state_mod.run_stop_checks(
+            rc = pipeline_mod.run_stop_checks(
                 Path("/nonexistent/path"),
                 "origin/main",
                 state_mod.StopCheckOptions(always_fetch=False, max_out=12000),
@@ -768,12 +778,12 @@ class TestNetsukeTargets:
         driver = driver_mod.BuildDriver("netsuke", "netsuke", "Netsukefile")
         with (
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "get_build_targets",
                 return_value=({"check-fmt", "lint", "markdownlint"}, None),
             ),
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "run_build_targets",
                 return_value={
                     "kind": "markdown",
@@ -784,7 +794,7 @@ class TestNetsukeTargets:
                 },
             ) as mock_run_targets,
         ):
-            rc = state_mod.evaluate_changes(state, REPO, 12000, driver)
+            rc = pipeline_mod.evaluate_changes(state, REPO, 12000, driver)
         assert rc == 0
         mock_run_targets.assert_called_once_with(
             REPO,
@@ -797,12 +807,12 @@ class TestNetsukeTargets:
         driver = driver_mod.BuildDriver("netsuke", "netsuke", "Netsukefile")
         with (
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "get_build_targets",
                 return_value=({"check-fmt", "lint", "markdownlint"}, None),
             ),
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "run_build_targets",
                 return_value={
                     "kind": "code",
@@ -813,7 +823,7 @@ class TestNetsukeTargets:
                 },
             ) as mock_run_targets,
         ):
-            rc = state_mod.evaluate_changes(state, REPO, 12000, driver)
+            rc = pipeline_mod.evaluate_changes(state, REPO, 12000, driver)
         assert rc == 0
         mock_run_targets.assert_called_once_with(
             REPO,
@@ -826,12 +836,12 @@ class TestNetsukeTargets:
         driver = driver_mod.BuildDriver("netsuke", "netsuke", "Netsukefile")
         with (
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "get_build_targets",
                 return_value=({"check-fmt", "lint", "typecheck"}, None),
             ),
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "run_build_targets",
                 return_value={
                     "kind": "code",
@@ -842,7 +852,7 @@ class TestNetsukeTargets:
                 },
             ) as mock_run_targets,
         ):
-            rc = state_mod.evaluate_changes(state, REPO, 12000, driver)
+            rc = pipeline_mod.evaluate_changes(state, REPO, 12000, driver)
         assert rc == 0
         mock_run_targets.assert_called_once_with(
             REPO,
@@ -859,12 +869,12 @@ class TestNetsukeTargets:
         driver = driver_mod.BuildDriver("netsuke", "netsuke", "Netsukefile")
         with (
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "get_build_targets",
                 return_value=({"markdownlint"}, None),
             ),
             mock.patch.object(
-                state_mod,
+                pipeline_mod,
                 "run_build_targets",
                 return_value={
                     "kind": "markdown",
@@ -875,7 +885,7 @@ class TestNetsukeTargets:
                 },
             ),
         ):
-            rc = state_mod.evaluate_changes(state, REPO, 12000, driver)
+            rc = pipeline_mod.evaluate_changes(state, REPO, 12000, driver)
         assert rc == exec_mod.BLOCKED_STATUS
         out = json.loads(capsys.readouterr().out)
         assert "Requested build targets: markdownlint" in out["reason"]
@@ -938,13 +948,13 @@ class TestDetectCategories:
     """Tests for detect_categories()."""
 
     def test_empty_file_list(self) -> None:
-        cats = state_mod.detect_categories([])
+        cats = formatting_mod.detect_categories([])
         assert all(v is False for v in cats.values()), (
             f"expected all False for empty list but got {cats}"
         )
 
     def test_py_file(self) -> None:
-        cats = state_mod.detect_categories(["src/app.py"])
+        cats = formatting_mod.detect_categories(["src/app.py"])
         assert cats["python_ts"] is True, (
             f"expected python_ts True for .py file but was {cats['python_ts']!r}"
         )
@@ -952,20 +962,20 @@ class TestDetectCategories:
         assert cats["markdown"] is False
 
     def test_ts_file(self) -> None:
-        cats = state_mod.detect_categories(["src/app.ts"])
+        cats = formatting_mod.detect_categories(["src/app.ts"])
         assert cats["python_ts"] is True, (
             f"expected python_ts True for .ts file but was {cats['python_ts']!r}"
         )
 
     def test_md_file(self) -> None:
-        cats = state_mod.detect_categories(["docs/readme.md"])
+        cats = formatting_mod.detect_categories(["docs/readme.md"])
         assert cats["markdown"] is True, (
             f"expected markdown True for .md file but was {cats['markdown']!r}"
         )
         assert cats["python_ts"] is False
 
     def test_mixed_py_and_md(self) -> None:
-        cats = state_mod.detect_categories(["src/app.py", "docs/readme.md"])
+        cats = formatting_mod.detect_categories(["src/app.py", "docs/readme.md"])
         assert cats["python_ts"] is True
         assert cats["markdown"] is True
         assert cats["rust"] is False
@@ -1139,7 +1149,7 @@ class TestFailState:
 
     def test_sets_ok_false_and_error(self) -> None:
         state = state_mod.HookState()
-        rc = state_mod.fail_state(state, "something went wrong")
+        rc = formatting_mod.fail_state(state, "something went wrong")
         assert state.ok is False, f"expected ok=False but was {state.ok!r}"
         assert state.error == "something went wrong", (
             f"expected error message but got {state.error!r}"
@@ -1148,7 +1158,7 @@ class TestFailState:
 
     def test_none_message(self) -> None:
         state = state_mod.HookState()
-        state_mod.fail_state(state, None)
+        formatting_mod.fail_state(state, None)
         assert state.ok is False
         assert state.error is None, f"expected error=None but got {state.error!r}"
 
@@ -1163,20 +1173,20 @@ class TestFormatReason:
 
     def test_returns_non_empty_string(self) -> None:
         state = state_mod.HookState()
-        reason = state_mod.format_reason(state)
+        reason = formatting_mod.format_reason(state)
         assert isinstance(reason, str), f"expected str but got {type(reason)}"
         assert len(reason) > 0, "expected non-empty reason string"
 
     def test_contains_base_ref(self) -> None:
         state = state_mod.HookState(base_ref="origin/main")
-        reason = state_mod.format_reason(state)
+        reason = formatting_mod.format_reason(state)
         assert "origin/main" in reason, (
             f"expected base_ref in reason but got {reason!r}"
         )
 
     def test_contains_error_message(self) -> None:
         state = state_mod.HookState(error="something broke")
-        reason = state_mod.format_reason(state)
+        reason = formatting_mod.format_reason(state)
         assert "something broke" in reason, (
             f"expected error message in reason but got {reason!r}"
         )
@@ -1192,7 +1202,7 @@ class TestBlockAndPrint:
 
     def test_writes_valid_json(self, capsys: pytest.CaptureFixture[str]) -> None:
         state = state_mod.HookState()
-        state_mod.block_and_print(state)
+        formatting_mod.block_and_print(state)
         out = json.loads(capsys.readouterr().out)
         assert isinstance(out, dict), f"expected JSON object but got {type(out)}"
 
@@ -1200,14 +1210,14 @@ class TestBlockAndPrint:
         self, capsys: pytest.CaptureFixture[str]
     ) -> None:
         state = state_mod.HookState()
-        state_mod.block_and_print(state)
+        formatting_mod.block_and_print(state)
         out = json.loads(capsys.readouterr().out)
         assert "decision" in out, f"expected 'decision' key in {out}"
         assert "reason" in out, f"expected 'reason' key in {out}"
 
     def test_decision_is_block(self, capsys: pytest.CaptureFixture[str]) -> None:
         state = state_mod.HookState()
-        state_mod.block_and_print(state)
+        formatting_mod.block_and_print(state)
         out = json.loads(capsys.readouterr().out)
         assert out["decision"] == "block", (
             f"expected block decision but got {out['decision']!r}"
@@ -1215,7 +1225,7 @@ class TestBlockAndPrint:
 
     def test_returns_zero(self) -> None:
         state = state_mod.HookState()
-        rc = state_mod.block_and_print(state)
+        rc = formatting_mod.block_and_print(state)
         assert rc == 0, f"expected return 0 but got {rc!r}"
